backend/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-experimental>=0.0.1
python-dotenv>=1.0.0
docx2txt>=0.8
pandas>=2.0.0
openpyxl>=3.1.2
python-docx>=0.8.11
pydub>=0.25.1
numpy>=1.24.0
faiss-cpu>=1.7.4
chromadb>=0.4.18 